Any keypress, or focus() call, will now trigger focus-visible.
This is a fairly big change.
I removed hadKeyboardEvent and made it so the polyfill now keeps tabs on
whether or not focus came from a pointing device.
It also keeps a reference to the clicked on element and uses that to
determine if clicking on one element calls focus() on another element.
In that scenario, we want focus-visible to be applied to the new element.

Fixes #88