Handle tabindex="-1"
Reach Router uses a focus wrapper element which is used to contain focus when no other element has it. It is used as a focus target when route is changed in SPA mode.

In general `tabindex="-1"` elements can only be focused without a keyboard so we should probably always ignore them.