hasKeyboardModality: suppress hadKeyboardEvent changes while typing.
While keydowns inside the input are suppressed, the method by which
you focused the input is remembered. So if you Tab into a text
field, type something, and hit Enter, then the app manually calls
focus() on the result: that other element _will_ have
the .focus-visible class, due to the _Tab_ key, not the Enter key.

Whereas if you click or tap into a text field, type something, and
hit Enter: when the app manually calls focus() on the result, it
will _not_ have the .focus-visible class, because the last relevant
interaction was clicking into the text field, not hitting Enter.